refactor: dispatch roots and staged plans move into their own namespaces
The DispatchRoots/ and StagedPlans/ folders now carry matching namespaces
(Stella.Ergosfare.Core.Abstractions.DispatchRoots / .StagedPlans) so the
root Core.Abstractions namespace stays uncluttered - maintainer's move,
plus the dependent wiring: the generator's availability probe metadata name
(without which emission silently degrades to no plans at all), the emitter's
fully-qualified type names in generated code, and the using directives in
consumers and tests.
