+++ src/Stella.Ergosfare.Core.Abstractions/DispatchRoots/IMessageResultRootVisitor.cs
@@ -1,5 +1,4 @@
-namespace Stella.Ergosfare.Core.Abstractions;
-
+namespace Stella.Ergosfare.Core.Abstractions.DispatchRoots;
 /// <summary>Generic re-entry point for consumers of <see cref="MessageResultRoot"/>.</summary>
 public interface IMessageResultRootVisitor<out TReturn, in TState>
 {
