+++ src/Stella.Ergosfare.Core.Abstractions/DispatchRoots/IMessageRootVisitor.cs
@@ -1,5 +1,4 @@
-namespace Stella.Ergosfare.Core.Abstractions;
-
+namespace Stella.Ergosfare.Core.Abstractions.DispatchRoots;
 /// <summary>Generic re-entry point for consumers of <see cref="MessageRoot"/>.</summary>
 public interface IMessageRootVisitor<out TReturn, in TState>
 {
