+++ src/Stella.Ergosfare.Core.Abstractions/DispatchRoots/IVoidPlanRootVisitor.cs
@@ -1,7 +1,6 @@
 using Stella.Ergosfare.Core.Abstractions.Handlers;
 
-namespace Stella.Ergosfare.Core.Abstractions;
-
+namespace Stella.Ergosfare.Core.Abstractions.DispatchRoots;
 /// <summary>Generic re-entry point for consumers of <see cref="VoidPlanRoot"/>.</summary>
 public interface IVoidPlanRootVisitor<out TReturn, in TState>
 {
