+++ src/Stella.Ergosfare.Core.Abstractions/DispatchRoots/VoidPlanRoot.cs
@@ -1,5 +1,4 @@
-namespace Stella.Ergosfare.Core.Abstractions;
-
+namespace Stella.Ergosfare.Core.Abstractions.DispatchRoots;
 /// <summary>
 /// A compile-time pipeline plan closed over a void message and its sole async handler;
 /// see <see cref="GeneratedDispatchRoots.AddVoidPlan{TMessage, THandler}"/> and
