+++ src/Stella.Ergosfare.Core.Abstractions/StagedPlans/IStagedResultPlanVisitor.cs
@@ -1,4 +1,4 @@
-namespace Stella.Ergosfare.Core.Abstractions;
+namespace Stella.Ergosfare.Core.Abstractions.StagedPlans;
 
 /// <summary>Generic re-entry point for consumers of <see cref="StagedResultPlan"/>.</summary>
 public interface IStagedResultPlanVisitor<out TReturn, in TState>
