+++ src/Stella.Ergosfare.Core/Internal/Mediator/StagedPlanGate.cs
@@ -1,5 +1,6 @@
 using Stella.Ergosfare.Core.Abstractions;
 using Stella.Ergosfare.Core.Abstractions.Registry.Descriptors;
+using Stella.Ergosfare.Core.Abstractions.StagedPlans;
 
 namespace Stella.Ergosfare.Core.Internal.Mediator;
